+++ eudi/openid4vp/openid4vp.go
@@ -1,17 +1,7 @@
 package openid4vp
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"net/http"
-	"net/url"
-	"strings"
-
 	"github.com/golang-jwt/jwt/v5"
-	"github.com/lestrrat-go/jwx/v3/jwk"
-	irma "github.com/privacybydesign/irmago"
 	"github.com/privacybydesign/irmago/eudi/openid4vp/dcql"
 )
 
@@ -176,16 +166,6 @@ type EncryptedResponsePayload struct {
 // VpToken is a map from dcql query id to a list of credentials (e.g. a list of sd-jwt vc's)
 type VpToken map[string][]string
 
-type AuthorizationResponseConfig struct {
-	State             string
-	QueryResponses    []dcql.QueryResponse
-	ResponseUri       string
-	ResponseType      string
-	ResponseMode      ResponseMode
-	CompatibilityMode CompatibilityMode
-	encryptionKey     *jwk.Key
-}
-
 // implement jwt.Claims interface, so we can decode the auth request JWT
 
 func (ar *AuthorizationRequest) GetExpirationTime() (*jwt.NumericDate, error) { return nil, nil }
@@ -203,175 +183,3 @@ const (
 	Compatibility_LatestDraft CompatibilityMode = "latest"
 	Compatibility_Draft24     CompatibilityMode = "draft24"
 )
-
-type OpenID4VPClient struct {
-	Compatibility CompatibilityMode
-	QueryHandlers []dcql.CredentialQueryHandler
-}
-
-func NewOpenID4VPClient(queryHandlers []dcql.CredentialQueryHandler) (*OpenID4VPClient, error) {
-	return &OpenID4VPClient {
-		Compatibility: Compatibility_Draft24,
-		QueryHandlers: queryHandlers,
-	}, nil
-}
-
-func (client *OpenID4VPClient) NewSession(fullUrl string) error {
-	components, err := url.Parse(fullUrl)
-
-	if err != nil {
-		return err
-	}
-
-	uri := components.Query().Get("request_uri")
-
-	irma.Logger.Infof("starting openid4vp session: %v\n", uri)
-	response, err := http.Get(uri)
-	if err != nil {
-		return err
-	}
-
-	defer response.Body.Close()
-
-	jawd, err := io.ReadAll(response.Body)
-
-	if err != nil {
-		return err
-	}
-
-	request, err := ParseAuthorizationRequestJwt(string(jawd))
-	if err != nil {
-		return err
-	}
-	irma.Logger.Infof("auth request: %#v\n", request)
-	return client.HandleAuthorizationRequest(request)
-}
-
-func (client *OpenID4VPClient) HandleAuthorizationRequest(request *AuthorizationRequest) error {
-	queryResponses, err := dcql.QueryCredentials(request.DcqlQuery, client.QueryHandlers)
-
-	if err != nil {
-		return err
-	}
-
-	httpClient := http.Client{}
-	authResponse := AuthorizationResponseConfig{
-		CompatibilityMode: client.Compatibility,
-		State:             request.State,
-		QueryResponses:    queryResponses,
-		ResponseUri:       request.ResponseUri,
-		ResponseType:      request.ResponseType,
-		ResponseMode:      request.ResponseMode,
-	}
-	responseReq, err := createAuthorizationResponseHttpRequest(authResponse)
-	if err != nil {
-		return err
-	}
-
-	response, err := httpClient.Do(responseReq)
-	if err != nil {
-		return err
-	}
-	if response.StatusCode != http.StatusOK {
-		return fmt.Errorf("response status was not ok: %v", response)
-	}
-	return nil
-}
-
-func ParseAuthorizationRequestJwt(authReqJwt string) (*AuthorizationRequest, error) {
-	parser := jwt.NewParser()
-	token, _, err := parser.ParseUnverified(string(authReqJwt), &AuthorizationRequest{})
-
-	typ, ok := token.Header["typ"]
-	if !ok {
-		return nil, errors.New("auth request JWT needs to contain 'typ' in header, but doesn't")
-	}
-	if typ != AuthRequestJwtTyp {
-		return nil, fmt.Errorf("auth request JWT typ in header should be %v but was %v", AuthRequestJwtTyp, typ)
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	claims := token.Claims.(*AuthorizationRequest)
-
-	return claims, nil
-}
-
-func createAuthorizationResponseHttpRequest(config AuthorizationResponseConfig) (*http.Request, error) {
-	values := url.Values{}
-
-	if config.ResponseMode == ResponseMode_DirectPost {
-		vpToken, err := createDirectPostVpToken(config.CompatibilityMode, config.QueryResponses)
-		if err != nil {
-			return nil, err
-		}
-		values.Add("vp_token", vpToken)
-	}
-
-	if config.ResponseMode == ResponseMode_DirectPostJwt {
-		if config.encryptionKey == nil {
-			return nil, fmt.Errorf("using response mode %v, but the encryption key is nil", ResponseMode_DirectPostJwt)
-		}
-		jwe, err := createDirectPostJwtEncryptedResponse(config.CompatibilityMode, config.QueryResponses, *config.encryptionKey)
-		if err != nil {
-			return nil, err
-		}
-		values.Add("response", jwe)
-	}
-
-	values.Add("state", config.State)
-
-	req, err := http.NewRequest(http.MethodPost, config.ResponseUri, strings.NewReader(values.Encode()))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-	req.Header.Add("Accept", "application/json")
-
-	return req, nil
-}
-
-func createDirectPostJwtEncryptedResponse(compatibility CompatibilityMode, queryResponses []dcql.QueryResponse, encryptionKey jwk.Key) (string, error) {
-	vpToken, err := createVpToken(compatibility, queryResponses)
-	if err != nil {
-		return "", err
-	}
-	payload := map[string]any{
-		"vp_token": vpToken,
-	}
-	return encryptJwe(payload, encryptionKey)
-}
-
-func encryptJwe(payload map[string]any, key jwk.Key) (string, error) {
-	return "", nil
-}
-
-func createVpToken(compatibility CompatibilityMode, queryResponses []dcql.QueryResponse) (any, error) {
-	if compatibility == Compatibility_LatestDraft {
-		content := map[string][]string{}
-		for _, resp := range queryResponses {
-			content[resp.QueryId] = resp.Credentials
-		}
-
-		return content, nil
-	}
-	if compatibility == Compatibility_Draft24 {
-		content := map[string]string{}
-		for _, resp := range queryResponses {
-			content[resp.QueryId] = resp.Credentials[0]
-		}
-		return content, nil
-	}
-	return nil, fmt.Errorf("%v is not a supported value for compatibility mode", compatibility)
-}
-
-func createDirectPostVpToken(compatibility CompatibilityMode, queryResponses []dcql.QueryResponse) (string, error) {
-	content, err := createVpToken(compatibility, queryResponses)
-	if err != nil {
-		return "", err
-	}
-	result, err := json.Marshal(content)
-	return string(result), err
-}
